+++ CHANGELOG.rst
@@ -3,6 +3,10 @@ Release Notes
 
 Forthcoming
 -----------
+* ...
+
+2.0.12 (2020-03-10)
+-------------------
 * [idioms] the either_or pattern, designed to replace the Chooser, `#283 <https://github.com/splintered-reality/py_trees/pull/283>`_
 * [behaviours] TickCounter, a timer based on tree ticks, `#283 <https://github.com/splintered-reality/py_trees/pull/283>`_
 * [behaviours] CheckBlackboardVariableValues, logical checks across multiple values, `#283 <https://github.com/splintered-reality/py_trees/pull/283>`_